fix_code_format

+++ src/pika_client_conn.cc
@@ -119,9 +119,9 @@ std::shared_ptr<Cmd> PikaClientConn::DoCmd(const PikaCmdArgsType& argv, const st
 
   // Process Command
   c_ptr->Execute();
-  int64_t duration = pstd::NowMicros() - start_us;
+  uint64_t duration = pstd::NowMicros() - start_us;
   g_cmdstat_map[opt].cmd_count.fetch_add(1);
-  g_cmdstat_map[opt].cmd_time_consuming.fetch_add(static_cast<uint64_t>(duration));
+  g_cmdstat_map[opt].cmd_time_consuming.fetch_add(duration);
 
   if (g_pika_conf->slowlog_slower_than() >= 0) {
     ProcessSlowlog(argv, start_us, c_ptr->GetDoDuration());
